+++ src/components/link/link.css
@@ -1,0 +1,3 @@
+a {
+  text-decoration: none;
+}
