--- src/style.css
@@ -1,121 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-:root {
-  --orange: #d64937;
-  --silver: #dcdcdc;
-  --blue-dark: #1b1b32;
-  --puter: #2d2d2d;
-  --shadow-gray: #3d3d3d;
-  --davy-grey: #535353;
-  --steel-wool: #777;
-
-  font-family: sans-serif;
-  font-size: 10px;
-  color: var(--orange);
-  background-color: var(--puter);
-}
-
-html {
-  scroll-behavior: smooth;
-}
-
-.btn {
-  cursor: pointer;
-
-  height: 40px;
-  padding: 0 10px;
-  border: 0;
-  border-radius: 2px;
-
-  font-family: "Open Sans", sans-serif;
-  font-size: 1.6rem;
-  font-weight: 300;
-  color: var(--orange);
-  text-transform: capitalize;
-
-  background-color: var(--shadow-gray);
-
-  transition: 0.1s;
-}
-
-.login-form {
-  position: absolute;
-  top: 20vh;
-  left: calc(50vw - 150px);
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  width: 300px;
-  margin: 0 auto;
-  padding: 1%;
-  border: 1px solid;
-  border-radius: 5px;
-
-  background-color: var(--puter);
-}
-
-.fields {
-  margin-bottom: 4%;
-  padding: 16px;
-  border: 1px var(--orange) solid;
-}
-
-.field-container {
-  display: inline-flex;
-  align-items: center;
-
-  width: 100%;
-  margin: 2% 0;
-
-  text-transform: capitalize;
-}
-
-.field-container > :first-child {
-  width: 30%;
-  font-size: 1.6rem;
-}
-
-.field-container > :nth-child(2) {
-  width: 70%;
-}
-
-.input-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.field-input {
-  height: 30px;
-  padding: 1% 5%;
-  border: 0;
-
-  color: var(--orange);
-
-  background-color: var(--shadow-gray);
-}
-
-.login-btn {
-  width: 50%;
-  max-width: 200px;
-  height: 32px;
-  margin: 2%;
-}
-
-.info-btn {
-  width: 50%;
-  max-width: 200px;
-  height: 32px;
-  margin: 2%;
-}
-
-@media (hover: hover) {
-  .btn:hover {
-    background-color: var(--davy-grey);
-  }
-}
