+++ src/styles/style.css
@@ -1,0 +1,31 @@
+@import url("modern-normalize.css");
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+:root {
+  --orange: #d64937;
+  --silver: #dcdcdc;
+  --blue-dark: #1b1b32;
+  --puter: #2d2d2d;
+  --shadow-gray: #3d3d3d;
+  --davy-grey: #535353;
+  --steel-wool: #777;
+
+  scroll-behavior: smooth;
+
+  font-family: sans-serif;
+  font-size: 10px;
+  color: var(--orange);
+
+  background-color: var(--puter);
+}
+
+@media (hover: hover) {
+  button:hover {
+    background-color: var(--davy-grey);
+  }
+}
